fix: address BUI-146 PR review findings
