+++ docs/dev/specs/tui-transcript.md
@@ -131,28 +131,8 @@
 - Ctrl+C interrupt is turn-local: stop current model step and active tool process, keep app/session alive.
 - Interrupt injects detail-only developer-role control message `User interrupted you`.
 - Post-interrupt state returns idle with input ready.
-- Ctrl+C must always provide an exit path from the TUI. A stuck or inconsistent runtime activity state must not trap the operator in the UI; while an interrupt is pending, another Ctrl+C quits the TUI rather than waiting indefinitely for cleanup.
-- In daemon-backed modes, forced TUI exit closes only that client. It detaches the local runtime owner without shutting down the daemon or force-closing the shared runtime. In embedded in-process mode, process exit closes the embedded server, but the TUI still detaches local owner state before quitting.
-- Forced TUI exit uses owner-drop release semantics with a detach-only close policy, distinct from normal idle release. Even if the runtime is idle by the time the second Ctrl+C is handled, the forced-exit request drops this client owner without closing the shared runtime solely because of that keypress.
-- If forced local owner drop fails because the daemon/connection is broken or incompatible, the TUI still exits locally, surfaces a diagnostic, and must not fall back to any path that can force-close an active shared runtime.
-- Forced TUI exit preserves recoverable local input before shutdown in the structured draft/retry buffer, including active submitted text, queued messages, pending injected input, and reviewer buffers. It does not invent terminal runtime state while preserving those buffers.
-- If local TUI state says the runtime is busy while the authoritative server activity state is idle or unavailable, the central invariant helper controls behavior: global debug or invariant panic mode fails fast with diagnostic detail, and default diagnostic mode surfaces an observable local diagnostic before recovering to the server state so quit remains available.
-- Runtime interrupt returns the authoritative server activity snapshot. A no-op interrupt leaves the TUI idle instead of entering pending-interrupt state.
-- Runtime interrupt responses carry `RuntimeInterruptResponse{Version, Activity, InputReconciliation}`. The activity and input reconciliation facts come from one server read-model snapshot and share the same `ReadModelVersion`.
-- Awaiting prompt/approval is active runtime activity without the ordinary model/goal spinner. Prompt contents and answer controls come from the prompt read model.
-- Runtime activity carries the server active kind for exclusive work such as user turns, workflow turns, goal loops, compaction, user shell, background work, and runtime maintenance. Local compaction or reviewer UI state may decorate the display, but it must not become a separate busy/idle authority.
-- Ctrl+C on a focused prompt answers/cancels that prompt first. If no prompt remains and the runtime is still active, Ctrl+C follows the normal interrupt path; while interrupt is pending, another Ctrl+C exits locally.
-- Runtime activity, input reconciliation, main-view snapshots, interrupt responses, versioned runtime activity/reconciliation events on `SessionActivity`, and migrated prompt read-model stream events carry one server `ReadModelVersion` with epoch, per-session generation, and monotonic sequence. The TUI applies runtime UI payloads by this version so stale responses or events cannot regress busy/idle or input-commitment state.
-- The existing `SessionActivity` stream uses its raw event sequence for transcript/input-diagnostic replay and transports typed runtime activity/reconciliation payloads that carry `ReadModelVersion`. Raw `AfterSequence` is never derived from `ReadModelVersion`; raw sequence and raw run events cannot drive busy/idle, and raw gaps trigger main-view hydration/reset.
-- On attach or reconnect, the TUI hydrates main view for the base `ReadModelVersion`, subscribes/resumes `SessionActivity` with only the raw stream cursor, buffers typed activity/reconciliation events that arrive before hydration applies, then applies buffered/live events only when their epoch/generation matches and their sequence is newer than the hydrated/applied version. Epoch or generation mismatch triggers hydration/reset.
-- The TUI runtime attachment owns raw session and prompt subscriptions as one cleanup group. Normal exit, forced local owner drop, partial subscription setup failure, and connection cleanup stop those streams; BUI-146 does not add a second runtime UI subscription.
-- A new runtime read-model epoch or generation observed on a typed `SessionActivity` payload or RPC response triggers main-view hydration/reset before the TUI continues pending interrupt or input recovery.
-- Authoritative idle or unavailable activity from a runtime UI activity event, interrupt response, or main-view hydration clears a pending interrupt and releases runtime-busy projection even when the terminal raw run event was missed. This cleanup does not create a terminal run state.
-- A terminal raw `run_state_changed` event can run interrupt cleanup only when it matches the current interrupt token's non-optional active-step `RunID` and `StepID`. Raw terminal events with missing or stale active-step identity, or events arriving after newer runtime activity/reconciliation, successor start, generation hydration, or forced-exit entry, are diagnostics only.
-- Before typed reconciliation proves input fate, the first stuck-running fix clears pending interrupt from authoritative idle/unavailable activity without inventing input fate and preserves ordinary visible draft on forced exit. Structured hidden-buffer recovery moves ambiguous active submitted text, queued text, injected buffers, and reviewer buffers into `DraftRecoveryBuffer` before unlocking input; if that recovery snapshot cannot be captured, input remains locked with a diagnostic.
-- Snapshot-based interrupt recovery restores local input only with typed reconciliation for submit, queue, shell, or compact operation IDs. Main-view refresh sends pending operation refs and applies the returned input-reconciliation snapshot. Ambiguous stream-gap recovery preserves local text in a recoverable diagnostic path instead of duplicating or dropping it.
-- Ordinary runtime work resumes after interrupt only from explicit user text. Only interruption of a server-owned `goal_loop` step creates goal suspension; interrupting normal user, shell, compaction, workflow, or background work while a goal is active does not. A suspended active goal loop resumes after the next successful non-workflow user turn according to the goal runtime contract; workflow-active sessions keep the goal passive and do not start a separate goal loop.
-- Crash recovery restores from the server runtime read model. A reopened session with no live runtime is idle; recovery diagnostics do not mark runtime active.
+- Resume after interrupt requires explicit user text.
+- Crash recovery is bifurcated: mid-step crash resumes via interrupt flow; otherwise restore normal state.
 - Failed prompt-history navigation emits plain terminal BEL with no transient UI notification.
 
 ## Path Autocomplete
