+++ server/transport/gateway.go
@@ -330,7 +330,7 @@ func protocolError(err error) (int, string) {
 		return protocol.ErrCodeInternalError, "internal error"
 	}
 	message := strings.TrimSpace(err.Error())
-	if errors.Is(err, context.Canceled) {
+	if errors.Is(err, context.Canceled) || errors.Is(err, serverapi.ErrRuntimeOperationCanceled) {
 		if message == "" || message == context.Canceled.Error() {
 			message = canceledByClientMessage
 		}
