+++ shared/serverapi/runtime_errors.go
@@ -1,0 +1,5 @@
+package serverapi
+
+import "errors"
+
+var ErrRuntimeOperationCanceled = errors.New("runtime operation canceled before execution")
